test(hooks): execute the two claimed misses instead of asserting about them
Three surfaces and a contract table said collection does not catch a
missing fixture or two same-named test functions in one module. Nothing
ran either one. Review on PR #5319: the claims were held up by string
comparisons between surfaces that agree with each other, which is the
same shape as the basename collision that was published as a catch for
three commits and was never true under this repository's import mode.

Both now run the real command against a tree carrying that defect under
production addopts. The control is the second half of each case: append
a syntax error to the same tree and require it to block. Without that, a
clean exit could mean collection never reached the fixture, and the test
would pass for a tree pytest declined to look at.

The additions crossed the 500-line taste threshold, so the executed
probes moved to tests/validation/test_collection_probes.py along the seam
the contract module already draws: one module tests which commands get
built, the other runs them and reads the exit code, and they fail for
different reasons.

Refs #5315
